# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LocalSpot API
  version: 1.0.0-draft
  description: |
    Đặc tả API thiết kế trước (design-first) cho LocalSpot. Nguồn: plan-v1.md §6, docs/requirements.md,
    docs/analysis/use-cases.md, docs/design/database.md.

    **Quy ước**
    - Prefix `/api/v1`, JSON UTF-8, thời gian ISO-8601 UTC.
    - Lỗi theo RFC 7807 (`application/problem+json`, schema `Problem`); lỗi validate có `errors[]` theo từng trường.
      Phân loại *(chốt 2026-10-03)*: **400** `MALFORMED_REQUEST` = request không đọc được (JSON hỏng, sai kiểu / thiếu
      tham số — client gửi đúng kiểu thì không gặp, nên không liệt kê ở từng operation); **422** = đọc được nhưng sai luật;
      **409** `CONCURRENT_MODIFICATION` (optimistic lock) / `DUPLICATE_RESOURCE` (trùng UNIQUE); **500** `INTERNAL_ERROR`
      không lộ chi tiết, kèm `errorId` trùng với dòng log máy chủ. Mã lỗi đầy đủ: `ErrorCode.java`.
    - Danh sách dài phân trang **cursor**: `?cursor=&limit=` → `{ items, nextCursor }`; `nextCursor = null` là hết.
    - Xác thực: `Authorization: Bearer <accessToken>` (15 phút). Refresh token nằm trong cookie HttpOnly
      `refresh_token` (Path=/api/v1/auth, SameSite=Strict) — S1.
    - Phân quyền ghi trong `x-permission` của từng operation (khớp bảng `permissions`).
    - Nhóm đường dẫn theo vai trò: công khai / thành viên ở gốc, `/owner/*` chủ địa điểm, `/moderation/*` kiểm duyệt viên, `/admin/*` quản trị viên.
...
servers:
  - url: http://localhost:8080/api/v1
    description: Dev
...
security:
  - bearerAuth: []
paths:
...
  /places:
    get:
      tags: [Places]
      summary: Danh sách địa điểm đã duyệt + lọc + sắp xếp (FR-10)
      description: |
        *(chốt 2026-10-04)* Keyset theo thứ tự đã chọn, phụ thêm `id DESC` để đồng điểm không trùng / sót giữa các trang.
        Cursor gắn với `sort` — dùng cursor của thứ tự khác → 422 `INVALID_CURSOR`.
      security: []
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: categoryId, in: query, schema: { type: integer, format: int64 }, description: Gồm cả danh mục con }
        - { name: amenityIds, in: query, style: form, explode: false, schema: { type: array, items: { type: integer, format: int64 } }, description: Phải có đủ tất cả (AND) }
        - { name: priceMax, in: query, schema: { type: integer, minimum: 0 }, description: 'Ngân sách (VND): `priceMin ≤ priceMax`; địa điểm chưa khai giá bị loại' }
        - { name: minRating, in: query, schema: { type: number, minimum: 1, maximum: 5 }, description: Theo điểm Bayesian hiển thị }
        - { name: city, in: query, schema: { type: string }, description: 'Khớp đúng tên, không phân biệt hoa thường / dấu' }
        - { name: sort, in: query, schema: { type: string, enum: [SCORE, NEWEST, MOST_REVIEWED], default: SCORE } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PlaceSummaryPage' } } } }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Places]
      summary: Đề xuất địa điểm mới → PENDING (FR-15, UC11)
      description: |
        Slug sinh từ tên + thành phố, bỏ dấu (`pho-thin-ha-noi`), trùng thì thêm `-2`, `-3`… Thẻ trùng slug gộp làm một.
        Danh mục / tiện ích không tồn tại, `priceMin > priceMax`, khung giờ mở = đóng → 422 theo trường.
        `photos` (không bắt buộc): JPEG/PNG, ≤ 10 ảnh/request, ≤ 5 MB và ≤ 40 MP mỗi ảnh, tổng ≤ 30 ảnh/địa điểm. Ảnh hỏng →
        422 `INVALID_IMAGE` trường `photos[i]` (không tạo gì); quá số ảnh → 422 `PHOTO_LIMIT_EXCEEDED`; request > 55 MB → 413;
        kho ảnh lỗi → 503. Ảnh trả về ở trạng thái `PROCESSING`, xử lý nền qua RabbitMQ *(E2, 2026-10-06)*.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [place]
              properties:
                place: { $ref: '#/components/schemas/PlaceCreateRequest' }
                photos: { type: array, maxItems: 10, items: { type: string, format: binary }, description: JPEG/PNG, ≤ 5 MB mỗi ảnh }
            encoding:
              place: { contentType: application/json }
      responses:
        '201':
          description: 'Đã tạo, chờ duyệt'
          headers:
            Location: { schema: { type: string }, description: '/api/v1/places/{slug}' }
          content: { application/json: { schema: { $ref: '#/components/schemas/PlaceDetail' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /places/{slug}:
    get:
      tags: [Places]
      summary: Chi tiết địa điểm theo slug (FR-13) — ghi nhận lượt xem theo ngày
      description: |
        Địa điểm chưa duyệt chỉ người đề xuất, chủ và kiểm duyệt viên xem được — người khác nhận 404 như không tồn tại.
        Lượt xem (+1 / request, gộp theo ngày giờ Việt Nam) chỉ tính cho địa điểm APPROVED và không tính khi chủ tự xem.
        Gửi kèm token (không bắt buộc) để nhận `myReviewId`, `claimable` theo người xem.
      security: []
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PlaceDetail' } } } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────────── Search (UC08) ─────────────────────────────
...
  /collections/{collectionId}/places/{placeId}:
    put:
      tags: [Collections]
      summary: Thêm địa điểm vào bộ sưu tập — idempotent
      parameters: [ { $ref: '#/components/parameters/CollectionId' }, { $ref: '#/components/parameters/PlaceId' } ]
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { note: { type: string, maxLength: 500 } } }
      responses:
        '204': { description: OK }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [Collections]
      summary: Bỏ địa điểm khỏi bộ sưu tập — idempotent
      parameters: [ { $ref: '#/components/parameters/CollectionId' }, { $ref: '#/components/parameters/PlaceId' } ]
      responses:
        '204': { description: OK }
...
  /owner/places/{placeId}:
    patch:
      tags: [Owner]
      summary: Chủ cập nhật thông tin địa điểm (FR-32)
      x-permission: place:update-own
      parameters: [ { $ref: '#/components/parameters/PlaceId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceUpdateRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PlaceDetail' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: '`version` lệch (`CONCURRENT_MODIFICATION`)' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /owner/places/{placeId}/photos:
    post:
      tags: [Owner]
      summary: Chủ thêm ảnh địa điểm
      description: |
        Cùng quy tắc ảnh như `POST /places` (JPEG/PNG, ≤ 10 ảnh/request, tổng ≤ 30 ảnh/địa điểm — ảnh `FAILED` không tính).
        Ảnh đầu tiên của địa điểm chưa có bìa thành ảnh bìa.
      x-permission: place:update-own
      parameters: [ { $ref: '#/components/parameters/PlaceId' } ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photos]
              properties:
                photos: { type: array, maxItems: 10, items: { type: string, format: binary } }
      responses:
        '201':
          description: 'Đã nhận, đang xử lý nền — các ảnh vừa thêm, `status: PROCESSING`'
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Photo' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { description: Request vượt 55 MB / ảnh vượt 5 MB (`PAYLOAD_TOO_LARGE`) }
        '422': { description: '`INVALID_IMAGE` (theo trường `photos[i]`) / `PHOTO_LIMIT_EXCEEDED`', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '503': { description: Kho ảnh không khả dụng (`SERVICE_UNAVAILABLE`) }
  /owner/places/{placeId}/photos/{photoId}:
    delete:
      tags: [Owner]
      summary: Chủ xóa ảnh
      description: Xóa bản ghi và các object trên kho (sau commit). Xóa ảnh bìa → ảnh chưa lỗi kế tiếp thành bìa. Ảnh không thuộc địa điểm → 404 `PHOTO_NOT_FOUND`.
      x-permission: place:update-own
      parameters: [ { $ref: '#/components/parameters/PlaceId' }, { name: photoId, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses:
        '204': { description: OK }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string }, description: Cursor mờ (opaque) từ `nextCursor` }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
    PlaceId: { name: placeId, in: path, required: true, schema: { type: integer, format: int64 } }
...
    CollectionId: { name: collectionId, in: path, required: true, schema: { type: integer, format: int64 } }
...
  responses:
...
    ValidationError:
      description: Dữ liệu không hợp lệ
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Chưa xác thực / token không hợp lệ
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: Không có quyền
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Không tìm thấy
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 ProblemDetail (Spring 6)
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: 'Mã lỗi ổn định cho frontend, ví dụ REVIEW_ALREADY_EXISTS, EMAIL_NOT_VERIFIED' }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        errorId:
          type: string
          format: uuid
          description: Chỉ có ở lỗi 5xx — mã tra cứu trùng với dòng log máy chủ (người dùng gửi mã này khi báo lỗi)
...
    PlaceStatus: { type: string, enum: [PENDING, APPROVED, REJECTED, HIDDEN] }
...
    PhotoStatus: { type: string, enum: [PROCESSING, READY, FAILED] }
...
    Photo:
      type: object
      description: |
        URL chỉ có khi `READY` (bucket đọc công khai, URL ổn định — chốt 2026-10-06): thumb ≤ 320 px, medium ≤ 960 px,
        large ≤ 1920 px theo cạnh dài, JPEG, không EXIF. Người ngoài chỉ thấy ảnh `READY`; người đề xuất / chủ / kiểm duyệt
        viên thấy cả `PROCESSING` / `FAILED`.
      properties:
        id: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/PhotoStatus' }
        thumbUrl: { type: [string, 'null'], format: uri }
        mediumUrl: { type: [string, 'null'], format: uri }
        largeUrl: { type: [string, 'null'], format: uri }
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }

    # ── Auth / user ──
...
    UserSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        displayName: { type: string }
        avatarUrl: { type: [string, 'null'], format: uri }
...
    Amenity:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        slug: { type: string }
        icon: { type: [string, 'null'] }
...
    OpeningHour:
      type: object
      required: [dayOfWeek, openTime, closeTime]
      properties:
        dayOfWeek: { type: integer, minimum: 1, maximum: 7, description: 1 = Thứ Hai }
        openTime: { type: string, pattern: '^\d{2}:\d{2}$' }
        closeTime: { type: string, pattern: '^\d{2}:\d{2}$', description: Nhỏ hơn openTime = mở qua đêm }
    PlaceSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        slug: { type: string }
        name: { type: string }
        category: { $ref: '#/components/schemas/CategoryRef' }
        address: { type: string }
        location: { $ref: '#/components/schemas/GeoPoint' }
        coverPhoto: { oneOf: [ { $ref: '#/components/schemas/Photo' }, { type: 'null' } ] }
        priceMin: { type: [integer, 'null'] }
        priceMax: { type: [integer, 'null'] }
        bayesianScore: { type: number, description: Điểm xếp hạng hiển thị }
        avgRating: { type: number, description: Trung bình thô }
        reviewCount: { type: integer }
        status: { $ref: '#/components/schemas/PlaceStatus' }
        distanceM: { type: [integer, 'null'], description: Chỉ có ở /places/nearby }
    CategoryRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        slug: { type: string }
    PlaceDetail:
      allOf:
        - { $ref: '#/components/schemas/PlaceSummary' }
        - type: object
          properties:
            description: { type: [string, 'null'] }
            city: { type: string }
            phone: { type: [string, 'null'] }
            website: { type: [string, 'null'] }
            photos: { type: array, items: { $ref: '#/components/schemas/Photo' } }
            openingHours: { type: array, items: { $ref: '#/components/schemas/OpeningHour' } }
            amenities: { type: array, items: { $ref: '#/components/schemas/Amenity' } }
            tags: { type: array, items: { type: string } }
            ratingDistribution:
              type: object
              description: Số review PUBLISHED theo sao
              properties:
                '1': { type: integer }
                '2': { type: integer }
                '3': { type: integer }
                '4': { type: integer }
                '5': { type: integer }
            checkinCount: { type: integer }
            owner: { oneOf: [ { $ref: '#/components/schemas/UserSummary' }, { type: 'null' } ] }
            claimable: { type: boolean, description: Chưa có chủ và người xem chưa có yêu cầu đang chờ }
            myReviewId: { type: [integer, 'null'], format: int64, description: 'Review của người xem (kể cả đã xóa → nút "Viết đánh giá" bị ẩn, D2)' }
            rejectReason: { type: [string, 'null'], description: Chỉ người đề xuất / kiểm duyệt viên thấy }
            version: { type: integer, description: 'Optimistic lock — gửi lại khi sửa (owner PATCH) / ra quyết định (2026-10-04)' }
    PlaceCreateRequest:
      type: object
      required: [name, categoryId, address, city, location]
      properties:
        name: { type: string, minLength: 2, maxLength: 200 }
        categoryId: { type: integer, format: int64 }
        description: { type: string, maxLength: 5000 }
        address: { type: string, maxLength: 300 }
        city: { type: string, maxLength: 100 }
        location: { $ref: '#/components/schemas/GeoPoint' }
        priceMin: { type: integer, minimum: 0 }
        priceMax: { type: integer, minimum: 0 }
        phone: { type: string, maxLength: 20 }
        website: { type: string, format: uri, maxLength: 300 }
        openingHours: { type: array, items: { $ref: '#/components/schemas/OpeningHour' } }
        amenityIds: { type: array, items: { type: integer, format: int64 } }
        tags: { type: array, maxItems: 10, items: { type: string, maxLength: 50 } }
    PlaceUpdateRequest:
      description: |
        Chủ không đổi được vị trí và danh mục (tránh "dời" địa điểm để thoát review) — muốn đổi phải báo cáo WRONG_INFO để kiểm duyệt viên xử lý.
        Ngữ nghĩa PATCH *(2026-10-04)*: trường vắng / null = giữ nguyên; chuỗi rỗng xóa `description` / `phone` / `website`;
        `openingHours`, `amenityIds` gửi lên là thay toàn bộ (mảng rỗng = xóa hết).
      type: object
      properties:
        description: { type: string, maxLength: 5000 }
        address: { type: string, maxLength: 300 }
        priceMin: { type: integer, minimum: 0 }
        priceMax: { type: integer, minimum: 0 }
        phone: { type: string, maxLength: 20 }
        website: { type: string, format: uri, maxLength: 300 }
        openingHours: { type: array, items: { $ref: '#/components/schemas/OpeningHour' } }
        amenityIds: { type: array, items: { type: integer, format: int64 } }
        version: { type: integer, description: Optimistic lock }
    PlaceSummaryPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PlaceSummary' } }
        nextCursor: { type: [string, 'null'] }
...